requirements.txt
python-documentcloud==3.3.4
requests
google-cloud-language
